<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰    
v    Before smashing the submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
v    If your PR doesn't close an issue, that's OK!  Just remove the Closes: #XXX line!
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

Closes: #6045

## What is the purpose of the change

To raise the minimum spot price in CL, we need to implement a MonotonicSqrt function on osmomath.BigDec to support higher precision.

See design document: https://app.clickup.com/37420681/v/dc/13nzm9-22453/13nzm9-45853

## Testing and Verifying

Copied tests from regular MonotonicSqrt.

## Benchmark

```
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/osmosis/osmomath
cpu: 12th Gen Intel(R) Core(TM) i7-1260P
                        │  reg.bench   │     big.bench      │
                        │    sec/op    │   sec/op     vs base   │
MonotonicSqrt-16          386.8µ ± 19%
MonotonicSqrt_BigDec-16                  470.0µ ± 2%
geomean                   386.8µ         470.0µ       ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean
```

## Documentation and Release Note

  - [ ] Does this pull request introduce a new feature or user-facing behavior changes?
  - [ ] Changelog entry added to `Unreleased` section of `CHANGELOG.md`?

Where is the change documented? 
  - [ ] Specification (`x/{module}/README.md`)
  - [ ] Osmosis documentation site
  - [ ] Code comments?
  - [ ] N/A